feat: expose default values behind flag
