+++ src/interfaces/class-transformer-options.interface.ts
@@ -57,4 +57,10 @@ export interface ClassTransformOptions {
    * DEFAULT: `false`
    */
   enableImplicitConversion?: boolean;
+
+  /**
+   * If set to true then class transformer will take default values for unprovided fields.
+   * This is useful when you convert a plain object to a class and have an optional field with a default value.
+   */
+  exposeDefaultValues?: boolean;
 }
