+++ test/functional/basic-functionality.spec.ts
@@ -1819,47 +1819,4 @@ describe('basic functionality', () => {
 
     expect(transformedClass.usersUndefined).toBeUndefined();
   });
-
-  it('should set default value if nothing provided', () => {
-    defaultMetadataStorage.clear();
-
-    class User {
-      @Expose({ name: 'AGE' })
-      @Transform(value => parseInt(value, 10))
-      age: number;
-
-      @Expose({ name: 'AGE_WITH_DEFAULT' })
-      @Transform(value => parseInt(value, 10))
-      ageWithDefault: number = 18;
-
-      @Expose({ name: 'FIRST_NAME' })
-      firstName: string;
-
-      @Expose({ name: 'FIRST_NAME_WITH_DEFAULT' })
-      firstNameWithDefault: string = 'default first name';
-
-      @Transform(value => !!value)
-      admin: boolean;
-
-      @Transform(value => !!value)
-      adminWithDefault: boolean = false;
-
-      lastName: string;
-
-      lastNameWithDefault: string = 'default last name';
-    }
-
-    const fromPlainUser = {};
-    const transformedUser = plainToClass(User, fromPlainUser);
-
-    expect(transformedUser).toBeInstanceOf(User);
-    expect(transformedUser).toEqual({
-      age: undefined,
-      ageWithDefault: 18,
-      firstName: undefined,
-      firstNameWithDefault: 'default first name',
-      adminWithDefault: false,
-      lastNameWithDefault: 'default last name',
-    });
-  });
 });
